AttachmentSheet, ui/utils: PR feedback
