+++ packages/app/ui/utils/clipboardUtils.ts
@@ -1,4 +1,5 @@
 import Clipboard from '@react-native-clipboard/clipboard';
+import * as ImagePicker from 'expo-image-picker';
 
 export const tryGetImageWithFormat = async (
   getter: () => Promise<string>,
@@ -31,4 +32,30 @@ export const getClipboardImageWithFallbacks = async (): Promise<{
   }
 
   return null;
+};
+
+export const createImageAssetFromClipboardData = (
+  clipboardData: { data: string; mimeType: string }
+): ImagePicker.ImagePickerAsset => {
+  const { data: imageData, mimeType } = clipboardData;
+  
+  const uri = imageData.startsWith('data:')
+    ? imageData
+    : `data:${mimeType};base64,${imageData}`;
+
+  return {
+    assetId: `clipboard-${Date.now()}`,
+    uri,
+    width: 300,
+    height: 300,
+    fileName:
+      mimeType === 'image/jpeg'
+        ? 'clipboard-image.jpg'
+        : 'clipboard-image.png',
+    fileSize: 0,
+    type: 'image',
+    duration: undefined,
+    exif: undefined,
+    base64: undefined,
+  };
 };
